+++ Editor/Steps/Processing/PackagesInstallerStep.cs
@@ -55,7 +55,7 @@ namespace Nomnom.UnityProjectPatcher.Editor.Steps {
                     return UniTask.FromResult(StepResult.Success);
                 }
                 
-                EditorUtility.DisplayDialog("Installing packages", $"The following packages are missing: {string.Join(", ", missingPackages)}", "OK");
+                EditorUtility.DisplayDialog("Installing packages", $"Press \"OK\" to install the following packages: {string.Join(", ", missingPackages)}", "OK");
                 EditorUtility.DisplayProgressBar("Installing packages", $"Installing {missingPackages.Length} package{(missingPackages.Length == 1 ? string.Empty : "s")}", 0.5f);
 
 #if UNITY_2020_3_OR_NEWER